requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Environment
python-decouple>=3.8,<4.0

# Image Processing
Pillow>=10.0,<11.0

# WSGI Server
gunicorn>=21.0,<22.0

# Redis (for caching and Celery)
redis>=5.0,<6.0
django-redis>=5.4,<6.0

# Celery (background tasks)
celery>=5.3,<6.0

# Payment Integration
mercadopago>=2.2,<3.0

# Email
django-anymail>=10.0,<11.0

# Storage
boto3>=1.28,<2.0
django-storages>=1.14,<2.0

# Development
django-debug-toolbar>=4.2,<5.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
coverage>=7.3,<8.0

# Code Quality
flake8>=6.1,<7.0
black>=23.0,<24.0
isort>=5.12,<6.0

# Documentation
drf-yasg>=1.21,<2.0
